+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## (3.2.0 IN PROGRESS)
 
+* New permission: Settings (Gobi Integration): Can view only. Refs UIGS-25.
+
 ## [3.1.0](https://github.com/folio-org/ui-gobi-settings/tree/v3.1.0) (2024-03-19)
 [Full Changelog](https://github.com/folio-org/ui-gobi-settings/compare/v3.0.0...v3.1.0)